net/mlx5: avoid implicit conversion to 64 bits
When compiling with MSVC, errors like the one below pop up:

../drivers/crypto/mlx5/mlx5_crypto_xts.c(488): warning C4334:
    '<<': result of 32-bit shift implicitly converted to 64 bits
    (was 64-bit shift intended?)

Depending on the situation, the fix is to do a 64-bit shift, or
be explicit about the type conversion by adding a cast.

Acked-by: Dariusz Sosnowski <dsosnowski@nvidia.com>

+++ drivers/crypto/mlx5/mlx5_crypto_xts.c
@@ -485,7 +485,7 @@ mlx5_crypto_xts_queue_pair_setup(struct rte_cryptodev *dev, uint16_t qp_id,
 	alloc_size = RTE_ALIGN(alloc_size, RTE_CACHE_LINE_SIZE);
 	alloc_size += (sizeof(struct rte_crypto_op *) +
 		       sizeof(struct mlx5_devx_obj *)) *
-		       RTE_BIT32(log_nb_desc);
+		       (size_t)RTE_BIT32(log_nb_desc);
 	qp = rte_zmalloc_socket(__func__, alloc_size, RTE_CACHE_LINE_SIZE,
 				socket_id);
 	if (qp == NULL) {
@@ -529,7 +529,7 @@ mlx5_crypto_xts_queue_pair_setup(struct rte_cryptodev *dev, uint16_t qp_id,
 		goto error;
 	qp->mkey = (struct mlx5_devx_obj **)RTE_ALIGN((uintptr_t)(qp + 1),
 							   RTE_CACHE_LINE_SIZE);
-	qp->ops = (struct rte_crypto_op **)(qp->mkey + RTE_BIT32(log_nb_desc));
+	qp->ops = (struct rte_crypto_op **)(qp->mkey + (size_t)RTE_BIT32(log_nb_desc));
 	qp->entries_n = 1 << log_nb_desc;
 	if (mlx5_crypto_indirect_mkeys_prepare(priv, qp, &mkey_attr,
 					       mlx5_crypto_gcm_mkey_klm_update)) {
